+++ VComponents/Components/Buttons/VLink/VWebLinkType.swift
@@ -1,5 +1,5 @@
 //
-//  VLinkType.swift
+//  VWebLinkType.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 2/7/21.
@@ -7,19 +7,19 @@
 
 import SwiftUI
 
-// MARK:- V Link Type
-typealias VLinkType = DerivedButtonType
+// MARK:- V Web Link Type
+typealias VWebLinkType = DerivedButtonType
 
-// MARK:- V Link Preset
+// MARK:- V Web Link Preset
 /// Enum that describes link preset, such as `primary`, `secondary`, `square`, or `plain`
 ///
 /// Custom type can be used via inits that do not take preset as a parameter
-public typealias VLinkPreset = DerivedButtonPreset
+public typealias VWebLinkPreset = DerivedButtonPreset
 
 // MARK:- Button
-extension VLinkType {
+extension VWebLinkType {
     @ViewBuilder static func linkButton<Content>(
-        buttonType: VLinkType,
+        buttonType: VWebLinkType,
         isEnabled: Bool,
         action: @escaping () -> Void,
         @ViewBuilder content: @escaping () -> Content