+++ VComponentsDemo/Scenes/Demos/Item Pickers/VMenuPickerDemoView.swift
@@ -104,7 +104,7 @@ private enum VMenuPickerButtonTypeHelper: Int, VPickableTitledItem {
     case plain
     case custom
     
-    var preset: VLinkPreset? {
+    var preset: VWebLinkPreset? {
         switch self {
         case .primary: return .primary()
         case .secondary: return .secondary()