+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Change Log
 
-#### [1.4.1(9)](https://github.com/VakhoKontridze/VComponents/releases/download/1.4.1/VComponents.xcframework.zip) — *2021 08 XX*
+#### [1.4.1(9)](https://github.com/VakhoKontridze/VComponents/releases/download/1.4.1/VComponents.xcframework.zip) — *2021 08 29*
 
 - `VLink` has been renamed to `VWebLink`
 - `VSectionListRowViewModelable` has been dropped for `Identifiable`